front/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dominate==2.7.0
Flask==2.2.2
Flask-Bootstrap4==4.0.2
Flask-CKEditor==0.4.6
Flask-WTF==1.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
requests==2.28.1
urllib3==1.26.12
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
